+++ hypervisor/arch/x86/guest/vlapic.c
@@ -31,7 +31,6 @@
 
 #include <hypervisor.h>
 
-#include "instr_emul_wrapper.h"
 #include "instr_emul.h"
 
 #include "vlapic_priv.h"
